✨ feat(vecs): register ND converters
