+++ src/coordinax/_src/vectors/dn/register_vectorapi.py
@@ -6,8 +6,114 @@ from typing import Any
 
 from plum import dispatch
 
+import quaxed.numpy as jnp
+from unxt.quantity import AbstractQuantity
+
 from .cartesian import CartesianAccND, CartesianPosND, CartesianVelND
 from .poincare import PoincarePolarVector
+
+###############################################################################
+
+
+@dispatch
+def vector(
+    cls: type[CartesianPosND] | type[CartesianVelND] | type[CartesianAccND],
+    x: AbstractQuantity,
+    /,
+) -> CartesianPosND | CartesianVelND | CartesianAccND:
+    """Construct an N-dimensional acceleration.
+
+    Examples
+    --------
+    >>> import unxt as u
+    >>> import coordinax as cx
+
+    1D vector:
+
+    >>> cx.vecs.CartesianPosND.from_(u.Quantity(1, "km"))
+    CartesianPosND(
+      q=Quantity[...](value=...i32[1], unit=Unit("km"))
+    )
+
+    >>> cx.vecs.CartesianPosND.from_(u.Quantity([1], "km"))
+    CartesianPosND(
+      q=Quantity[...](value=...i32[1], unit=Unit("km"))
+    )
+
+    >>> cx.vecs.CartesianVelND.from_(u.Quantity(1, "km/s"))
+    CartesianVelND(
+      d_q=Quantity[...]( value=...i32[1], unit=Unit("km / s") )
+    )
+
+    >>> cx.vecs.CartesianVelND.from_(u.Quantity([1], "km/s"))
+    CartesianVelND(
+      d_q=Quantity[...]( value=...i32[1], unit=Unit("km / s") )
+    )
+
+    >>> cx.vecs.CartesianAccND.from_(u.Quantity(1, "km/s2"))
+    CartesianAccND(
+      d2_q=Quantity[...]( value=...i32[1], unit=Unit("km / s2") )
+    )
+
+    >>> cx.vecs.CartesianAccND.from_(u.Quantity([1], "km/s2"))
+    CartesianAccND(
+      d2_q=Quantity[...]( value=...i32[1], unit=Unit("km / s2") )
+    )
+
+    2D vector:
+
+    >>> cx.vecs.CartesianPosND.from_(u.Quantity([1, 2], "km"))
+    CartesianPosND(
+      q=Quantity[...](value=...i32[2], unit=Unit("km"))
+    )
+
+    >>> cx.vecs.CartesianVelND.from_(u.Quantity([1, 2], "km/s"))
+    CartesianVelND(
+      d_q=Quantity[...]( value=...i32[2], unit=Unit("km / s") )
+    )
+
+    >>> cx.vecs.CartesianAccND.from_(u.Quantity([1, 2], "km/s2"))
+    CartesianAccND(
+      d2_q=Quantity[...]( value=...i32[2], unit=Unit("km / s2") )
+    )
+
+    3D vector:
+
+    >>> cx.vecs.CartesianPosND.from_(u.Quantity([1, 2, 3], "km"))
+    CartesianPosND(
+      q=Quantity[...](value=...i32[3], unit=Unit("km"))
+    )
+
+    >>> cx.vecs.CartesianVelND.from_(u.Quantity([1, 2, 3], "km/s"))
+    CartesianVelND(
+      d_q=Quantity[...]( value=...i32[3], unit=Unit("km / s") )
+    )
+
+    >>> cx.vecs.CartesianAccND.from_(u.Quantity([1, 2, 3], "km/s2"))
+    CartesianAccND(
+      d2_q=Quantity[...]( value=...i32[3], unit=Unit("km / s2") )
+    )
+
+    4D vector:
+
+    >>> cx.vecs.CartesianPosND.from_(u.Quantity([1, 2, 3, 4], "km"))
+    CartesianPosND(
+      q=Quantity[...](value=...i32[4], unit=Unit("km"))
+    )
+
+    >>> cx.vecs.CartesianVelND.from_(u.Quantity([1, 2, 3, 4], "km/s"))
+    CartesianVelND(
+      d_q=Quantity[...]( value=...i32[4], unit=Unit("km / s") )
+    )
+
+    >>> cx.vecs.CartesianAccND.from_(u.Quantity([1, 2, 3, 4], "km/s2"))
+    CartesianAccND(
+      d2_q=Quantity[...]( value=...i32[4], unit=Unit("km / s2") )
+    )
+
+    """
+    return cls(jnp.atleast_1d(x))
+
 
 ###############################################################################
 # Cartesian
